Update Default User Set 999999 employee id And Separate master controller with different controller.
